requirements.txt
Django==5.2
djangorestframework==3.16.0
Pillow==10.2.0 